Crate circuit_runtime_types

Structs

BlockHashCount

BlockLength

BlockWeights

We allow for 2 seconds of compute with a 6 second average block time.

SS58Prefix

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~10% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

CONTRACTS_DEBUG_OUTPUT

DAYS

EXISTENTIAL_DEPOSIT

The existential deposit. Set to 1/10 of the Connected Relay Chain.

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 of a second of compute with a 12 second average block time.

MICROUNIT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLIUNIT

MINUTES

NORMAL_DISPATCH_RATIO

SLOT_DURATION

UNIT

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

Amount

Balance of an account.

AssetId

Asset Id.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.